requirements.txt
python-dotenv
azure-ai-documentintelligence
streamlit
openai
pymongo